refactor(#1086): consolidate bodyUsesArguments imports to shared helper
The shared helper at src/codegen/helpers/body-uses-arguments.ts already
provides the iterative-DFS implementation with module-level WeakMap memo
cache (delivered via #1085 + initial extraction). This commit completes
the dedup pass by routing the remaining 5 call sites in
class-bodies.ts (2), declarations.ts (1), and literals.ts (2) directly
through the helper instead of via the function-body.ts re-export, and
removes the now-unused re-export from function-body.ts.

Net effect: collapses the post-#96 O(N²) re-walk to O(N) for files
with deep nested-function ASTs once any subtree has been walked, since
all call sites now share a single WeakMap<ts.Node, boolean> cache
keyed on stable ts.Node identity.

Verification:
- npx tsc --noEmit: clean
- tests/issue-1053-arguments-global-staleness.test.ts: 1/1 pass
- tests/equivalence/arguments-* + nested-class-declarations: same
  failure count vs main (3 pre-existing fails in
  arguments-nested-and-loops.test.ts, no new regressions)

[CHECKLIST-FOXTROT]
